+++ source/conf.py
@@ -110,7 +110,6 @@ exclude_patterns = [
     'user-manual/capabilities/malware-detection/**',
     'user-manual/capabilities/sec-config-assessment/**',
     'user-manual/capabilities/system-calls-monitoring/**',
-    'user-manual/capabilities/system-inventory/**',
     'user-manual/capabilities/vulnerability-detection/**',
     'user-manual/indexer-api/**',
     'user-manual/agent/agent-enrollment/deployment-variables/**',